requirements.txt
streamlit>=1.28.0
torch>=2.0.0
transformers>=4.50.0
Pillow>=9.0.0
requests>=2.28.0
accelerate>=0.20.0
numpy>=1.24.0
matplotlib>=3.6.0
plotly>=5.15.0
pandas>=1.5.0
loguru>=0.7.0
psutil>=5.9.0
nvidia-ml-py>=12.535.133
